cleaning up the code and removing legacy items.
