+++ Conference-Android/app/src/main/AndroidManifest.xml
@@ -25,10 +25,6 @@
         android:logo="@drawable/icon"
         android:theme="@style/Theme.Googleio" >
         <activity
-            android:name=".ui.activities.MainActivity"
-            android:label="@string/app_name" >
-        </activity>
-        <activity
             android:name=".ui.activities.SplashActivity"
             android:label="@string/app_name" >
