+++ Conference-Android/app/src/main/java/com/sagetech/conference_android/app/ConferenceApplication.java
@@ -3,16 +3,9 @@ package com.sagetech.conference_android.app;
 import android.app.Application;
 import android.content.Context;
 
-import com.sagetech.conference_android.app.api.ConferenceController;
-import com.sagetech.conference_android.app.model.EventData;
 import com.sagetech.conference_android.app.util.ConferenceModule;
 
-import java.util.List;
-
-import javax.inject.Inject;
-
 import dagger.ObjectGraph;
-import rx.Observable;
 import timber.log.Timber;
 
 import static timber.log.Timber.DebugTree;
@@ -23,8 +16,6 @@ import static timber.log.Timber.HollowTree;
  */
 public class ConferenceApplication extends Application {
     private ObjectGraph objectGraph;
-    @Inject ConferenceController conferenceController;
-    @Inject Observable<List<EventData>> cachedGetEventsObservable;
 
     @Override
     public void onCreate() {
@@ -52,22 +43,4 @@ public class ConferenceApplication extends Application {
         return (ConferenceApplication) context.getApplicationContext();
     }
 
-    /* ============================================================ */
-    // LEGACY OPERATIONS
-
-    /**
-     * @deprecated
-     * @return
-     */
-    public ConferenceController getConferenceController() {
-        return conferenceController;
-    }
-
-    /**
-     * @deprecated
-     * @return
-     */
-    public Observable<List<EventData>> getCachedGetEventsObservable() {
-        return cachedGetEventsObservable;
-    }
 }
