+++ Conference-Android/app/src/main/java/com/sagetech/conference_android/app/ui/activities/ConferenceListActivity.java
@@ -14,6 +14,7 @@ import android.widget.TextView;
 import com.sagetech.conference_android.app.R;
 import com.sagetech.conference_android.app.api.ConferenceController;
 import com.sagetech.conference_android.app.model.ConferenceData;
+import com.sagetech.conference_android.app.ui.adapters.ConferencesAdapter;
 import com.sagetech.conference_android.app.ui.presenter.ConferenceListActivityPresenter;
 import com.sagetech.conference_android.app.ui.presenter.IConferenceListActivity;
 
@@ -25,7 +26,7 @@ import butterknife.ButterKnife;
 import butterknife.InjectView;
 import timber.log.Timber;
 
-public class ConferenceListActivity extends InjectableActionBarActivity implements IConferenceListActivity {
+public class ConferenceListActivity extends InjectableActionBarActivity implements IConferenceListActivity, ConferencesAdapter.ConferencesOnClickListener {
 
     @Inject
     ConferenceController conferenceController;
@@ -78,7 +79,7 @@ public class ConferenceListActivity extends InjectableActionBarActivity implemen
 
     @Override
     public void populateConferences(List<ConferenceData> datas) {
-        mAdapter = new ConferencesAdapter(datas);
+        mAdapter = new ConferencesAdapter(datas, this);
         mRecyclerView.setAdapter(mAdapter);
     }
 
@@ -90,66 +91,8 @@ public class ConferenceListActivity extends InjectableActionBarActivity implemen
         startActivity(conferenceSessionListIntent);
     }
 
-    public class ConferencesAdapter extends RecyclerView.Adapter<ConferencesAdapter.ViewHolder> {
-
-        List<ConferenceData> conferenceDatas;
-
-        public ConferencesAdapter(List<ConferenceData> conferenceDatas) {
-            this.conferenceDatas = conferenceDatas;
-        }
-
-        @Override
-        public ConferencesAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-            View v = LayoutInflater.from(parent.getContext())
-                    .inflate(R.layout.conference_list_view_item, parent, false);
-
-            ViewHolder vh = new ViewHolder(v);
-
-            return vh;
-        }
-
-        public ConferenceData getItem(int position) {
-            return conferenceDatas.get(position);
-        }
-
-        @Override
-        public void onBindViewHolder(ConferencesAdapter.ViewHolder holder, int position) {
-            holder.setConferenceData(getItem(position));
-        }
-
-        @Override
-        public int getItemCount() {
-            return conferenceDatas.size();
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return getItem(position).getId();
-        }
-
-        public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
-            @InjectView(R.id.name) public TextView nameView;
-            private ConferenceData conferenceData;
-
-            public ViewHolder(View view) {
-                super(view);
-                ButterKnife.inject(this, view);
-                view.setOnClickListener(this);
-            }
-
-            public void setName(String name) {
-                nameView.setText(name);
-            }
-
-            public void setConferenceData(ConferenceData conferenceData) {
-                this.conferenceData = conferenceData;
-                setName(conferenceData.getName());
-            }
-
-            @Override
-            public void onClick(View v) {
-                launchConferenceSessionsListActivity(conferenceData.getId());
-            }
-        }
+    @Override
+    public void clicked(Integer conferenceId) {
+        launchConferenceSessionsListActivity(conferenceId);
     }
 }
