--- Conference-Android/app/src/main/java/com/sagetech/conference_android/app/ui/activities/MainActivity.java
@@ -1,22 +1,0 @@
-package com.sagetech.conference_android.app.ui.activities;
-
-import android.os.Bundle;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.ActionBarActivity;
-
-import com.sagetech.conference_android.app.R;
-
-
-public class MainActivity extends InjectableActionBarActivity {
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        final ActionBar actionBar = getSupportActionBar();
-        actionBar.setTitle(R.string.app_name);
-    }
-
-}
