+++ Conference-Android/app/src/main/java/com/sagetech/conference_android/app/ui/activities/SplashActivity.java
@@ -3,15 +3,10 @@ package com.sagetech.conference_android.app.ui.activities;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
-import android.util.Log;
 
 import com.sagetech.conference_android.app.R;
-import com.sagetech.conference_android.app.model.EventData;
 
-import java.util.List;
 import java.util.concurrent.TimeUnit;
-
-import javax.inject.Inject;
 
 import rx.Observable;
 import rx.Subscriber;
@@ -19,14 +14,11 @@ import rx.Subscription;
 import rx.android.app.AppObservable;
 import rx.android.schedulers.AndroidSchedulers;
 import rx.schedulers.Schedulers;
+import timber.log.Timber;
 
 public class SplashActivity extends InjectableActionBarActivity {
 
-    private static final String TAG = "SplashActivity";
     private Subscription subscription;
-
-    @Inject
-    Observable<List<EventData>> cachedGetEventsObservable;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -37,14 +29,15 @@ public class SplashActivity extends InjectableActionBarActivity {
         final ActionBar actionBar = getSupportActionBar();
         actionBar.hide();
 
-        subscription = AppObservable.bindActivity(this, cachedGetEventsObservable)
+
+        subscription = AppObservable.bindActivity(this, Observable.just(1))
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribeOn(Schedulers.io())
-                .delay(5, TimeUnit.SECONDS)
-                .subscribe(new Subscriber<List<EventData>>() {
+                .delay(3, TimeUnit.SECONDS)
+                .subscribe(new Subscriber<Integer>() {
                     @Override
                     public void onCompleted() {
-                        Log.i(TAG, "onCompleted");
+                        Timber.i("onCompleted");
                         Intent intent = new Intent(getApplicationContext(), ConferenceListActivity.class);
                         startActivity(intent);
                         finish();
@@ -52,15 +45,15 @@ public class SplashActivity extends InjectableActionBarActivity {
 
                     @Override
                     public void onError(Throwable e) {
-                        Log.i(TAG, "onError", e);
+                        Timber.i("onError", e);
                     }
 
                     @Override
-                    public void onNext(List<EventData> s) {
-                        Log.i(TAG, "On Next");
+                    public void onNext(Integer s) {
+                        Timber.i("On Next");
                     }
                 });
-        Log.i(TAG, "done!!!");
+        Timber.i("done!!!");
 
     }
