+++ Conference-Android/app/src/main/java/com/sagetech/conference_android/app/ui/presenter/ConferenceSessionListActivityPresenter.java
@@ -17,12 +17,12 @@ import timber.log.Timber;
  * Created by adam on 2/21/15.
  */
 public class ConferenceSessionListActivityPresenter {
-    private ConferenceSessionListActivity conferenceSessionListActivity;
+    private IConferenceSessionActivity conferenceSessionListActivity;
     private ConferenceController conferenceController;
     private Long conferenceId;
     private Subscription subscription;
 
-    public ConferenceSessionListActivityPresenter(ConferenceSessionListActivity conferenceSessionListActivity, ConferenceController conferenceController, Long conferenceId) {
+    public ConferenceSessionListActivityPresenter(IConferenceSessionActivity conferenceSessionListActivity, ConferenceController conferenceController, Long conferenceId) {
         this.conferenceSessionListActivity = conferenceSessionListActivity;
         this.conferenceController = conferenceController;
         this.conferenceId = conferenceId;
