--- Conference-Android/app/src/main/res/layout/activity_main.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-android:orientation="horizontal"
-android:layout_width="match_parent"
-android:layout_height="match_parent">
-<fragment android:name="com.sagetech.conference_android.app.ui.fragments.FullScheduleFragment"
-    android:id="@+id/list"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent" />
-</LinearLayout>
